Map store to merchant account
When processing payments in international stores we want to
associate the payments with a merchant account. This allows
transactions isolated by store and makes the record keeping
process easier.

Previously, merchant accounts were mapped to a currency. However,
this does not work for business which operate in multiple countries
that share a currency (ie. the Euro zone). To make it possible to
have multiple merchant accounts per currency, we change the association
from currency to merchant account to store code to merchant account.